Add Data Grid - Populate Grid From Web Service
This is a new Data Grid created in Angular. It uses Bootstrap for
styling. The grid has a very basic layout.

The grid loads itself from a JSON file. This pattern is used to simulate
a Web Service call. It uses Angular's HttpClient to load response from a
JSON file.